Refactor: 게시판 Detail page
