+++ src/routes/Board/BoardDetail/index.tsx
@@ -1,5 +1,12 @@
 import { Params, useNavigate, useParams } from "react-router-dom";
-import { useBoardDetail } from "./../../../hooks/board/useBoardDetail";
+import { useBoardDetail } from "../../../hooks/board/useBoardDetail";
+import { useState } from "react";
+import { useBoardDelete } from "../../../hooks/board/useBoardDelete";
+import { waldregAxios as axios } from "../../../apis/axios";
+import { useCommentList } from "../../../hooks/board/comment/useCommentList";
+import React from "react";
+import useCurUser from "../../../hooks/curuser/useCurUser";
+import Container from "../../../components/common/container";
 import {
   BoardButton,
   BoardButtonBox,
@@ -10,30 +17,23 @@ import {
   BoardTitle,
   BoardTopBox,
 } from "./style";
-import { useBoardDelete } from "./../../../hooks/board/useBoardDelete";
-import FONT from "./../../../constants/fonts";
-import BoardCommentCreate from "./../BoardCommentCreate/index";
-import BoardCommentList from "./../BoardCommentList/index";
-import { useCommentList } from "../../../hooks/board/comment/useCommentList";
-import { useState } from "react";
-import Modal from "../../common/modal";
+import FONT from "../../../constants/fonts";
+import {
+  FileDetailBox,
+  FileDetailTitle,
+  FileListBox,
+} from "../../../components/board/BoardFileUpload/style";
+import { FileDownLoadIcon } from "../../../components/Icons/BoardIcons";
+import BoardCommentCreate from "../../../components/board/BoardCommentCreate";
+import BoardCommentList from "../../../components/board/BoardCommentList";
+import Modal from "../../../components/common/modal";
 import {
   CategoryDeleteContent,
   SettingButtonBox,
   SettingCancelButton,
   SettingSaveButton,
   SettingTitle,
-} from "../../../routes/Setting/Board/style";
-import {
-  FileDetailBox,
-  FileDetailTitle,
-  FileListBox,
-} from "../BoardFileUpload/style";
-import { FileDownLoadIcon } from "../../Icons/BoardIcons";
-import axios from "axios";
-import React from "react";
-import useCurUser from "./../../../hooks/curuser/useCurUser";
-import Container from "../../common/container";
+} from "../../Setting/Board/style";
 
 const BoardDetail = () => {
   const navigate = useNavigate();
@@ -138,6 +138,8 @@ const BoardDetail = () => {
           )}
       </BoardTopBox>
 
+      <BoardContent style={FONT.BODY1}>{replaceValue}</BoardContent>
+
       {files && (
         <FileListBox>
           {files.map((file, i) => {
@@ -153,8 +155,6 @@ const BoardDetail = () => {
           })}
         </FileListBox>
       )}
-
-      <BoardContent style={FONT.BODY1}>{replaceValue}</BoardContent>
 
       <BoardCommentCount style={FONT.SUBTITLE2}>
         {commentLists?.max_idx}개의 댓글
