+++ src/routes/Board/index.tsx
@@ -1,9 +1,9 @@
 import { Routes, Route } from "react-router-dom";
 import BoardCreate from "../../components/board/BoardCreate/index";
-import BoardDetail from "../../components/board/BoardDetail";
 import BoardLayout from "./BoardLayout/index";
 import BoardPage from "./BoardPage";
 import BoardUpdate from "./../../components/board/BoardUpdate/index";
+import BoardDetail from "./BoardDetail";
 
 const Board = () => {
   return (
